chore: replace Timer class in a separate class (possibly soon to be deleted)
