+++ src/main/kotlin/algorithms/layout/ForceAtlas2.kt
@@ -22,23 +22,6 @@ import kotlin.random.Random
 *   println(layout)
  */
 
-class Timer(private val name: String = "Timer") {
-    private var startTime: Long = 0
-    private var totalTime: Long = 0
-
-    fun start() {
-        startTime = System.currentTimeMillis()
-    }
-
-    fun stop() {
-        totalTime += (System.currentTimeMillis() - startTime)
-    }
-
-    fun display() {
-        println("$name took ${"%.2f".format(totalTime / 1000.0)} seconds")
-    }
-}
-
 class ForceAtlas2(
     private val outboundAttractionDistribution: Boolean = true,
     private val edgeWeightInfluence: Double = 1.0,
